+++ src/constants/columns.js
@@ -5,9 +5,9 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-const COLUMNS_ATTRS = [ 'name', 'nullable', 'default', 'type', 'collate' ]
-const COLUMN_CONSTRAINTS = [ 'unique', 'primaryKey', 'references' ]
-const ALL_PROPERTIES = [ ...COLUMNS_ATTRS, ...COLUMN_CONSTRAINTS, 'formerNames', 'force' ]
+const ATTRS = [ 'name', 'nullable', 'default', 'type', 'collate' ]
+const CONSTRAINTS = [ 'unique', 'primaryKey', 'references' ]
+const ALL_PROPERTIES = [ ...ATTRS, ...CONSTRAINTS, 'formerNames', 'force', 'autoIncrement' ]
 
 const DEFAULTS = {
   nullable: true,
@@ -15,11 +15,12 @@ const DEFAULTS = {
   force: false,
   collate: null,
   formerNames: null,
+  autoIncrement: false,
 }
 
 module.exports = {
   DEFAULTS,
-  ATTRS: COLUMNS_ATTRS,
-  CONSTRAINTS: COLUMN_CONSTRAINTS,
+  ATTRS,
+  CONSTRAINTS,
   ALL_PROPERTIES,
 }