+++ src/model/index.js
@@ -9,6 +9,7 @@ const R = require('ramda')
 const chalk = require('chalk')
 const Sql = require('../sql')
 const Seeds = require('./seeds')
+const Sequence = require('../sequence')
 const Logger = require('../logger')
 
 const utils = require('../utils')
@@ -34,6 +35,7 @@ module.exports = function (options) {
 
   let _dbColumns = null
   let _dbConstraints = null
+  let _sequences = null
 
   const _schema = _parseSchema(schema)
   const _table = _schema.table
@@ -51,6 +53,20 @@ module.exports = function (options) {
 
   if (_schema.seeds) {
     _seeds.add(_schema.seeds)
+  }
+
+  const sequenceColumns = _schema.columns.filter(R.prop('autoIncrement'))
+  if (sequenceColumns.length) {
+    _sequences = sequenceColumns.map((column) => {
+      const sequence = new Sequence({
+        client,
+        name: `${_tableName}_${column.name}`,
+        schema: _schemaName,
+        sequence: column.autoIncrement,
+      })
+      column.default = sequence.getSqlIncrement()
+      return sequence
+    })
   }
 
   const _getSchema = () => _schema
@@ -72,7 +88,7 @@ module.exports = function (options) {
       and n.nspname = ic.table_schema
     where t.relname = '${_tableName}'
       and n.nspname = '${_schemaName}';
-  `).then(R.prop('rows')).then(parser.dbColumns)
+  `).then(R.prop('rows')).then(parser.dbColumns(_schemaName))
     return _dbColumns
   }
 
@@ -340,7 +356,11 @@ module.exports = function (options) {
           `you need to set 'force: true' for '${column.name}' column`,
         )
     } else if (key === 'default') {
-      return Sql.create('set default', `${alterTable} alter column ${column.name} set default ${value};`)
+      if (utils.isExist(value)) {
+        return Sql.create('set default', `${alterTable} alter column ${column.name} set default ${value};`)
+      } else {
+        return Sql.create('set default', `${alterTable} alter column ${column.name} drop default;`)
+      }
     }
     return null
   }
@@ -382,11 +402,10 @@ module.exports = function (options) {
   }
 
   const _createTable = (force) => {
-    const sql = new Sql()
     const columns = _schema.columns
       .map(_getColumnDescription)
       .join(',\n\t')
-    return sql.add([
+    return new Sql([
       force ? Sql.create('drop table', `drop table if exists ${_table} cascade;`) : null,
       Sql.create('create table', `create table ${_table} (\n\t${columns}\n);`),
     ])
@@ -396,10 +415,7 @@ module.exports = function (options) {
     COLUMNS.ATTRS.reduce((acc, key) => {
       const dbValue = dbColumn[key]
       const schemaValue = column[key]
-      if (
-        utils.isExist(schemaValue) &&
-        String(dbValue) !== String(schemaValue)
-      ) {
+      if (String(dbValue) !== String(schemaValue)) {
         acc[key] = schemaValue
         switch (key) {
           case 'type': {
@@ -424,17 +440,36 @@ module.exports = function (options) {
     ))
     if (hasConstraints) {
       const inserts = _seeds.inserts()
-      return new Sql().add(inserts.map((insert) => Sql.create('insert seed', insert)))
+      return new Sql(inserts.map((insert) => Sql.create('insert seed', insert)))
     } else {
       logger.error(`To use seeds, you need to set at least one constraint (primaryKey || unique)`)
       return null
     }
   }
 
+  const _getSqlSequenceChangesFrom = R.when(
+    R.identity,
+    R.pipe(
+      R.map((seq) => seq.getChanges()),
+      (promises) => Promise.all(promises),
+      R.then(
+        R.pipe(
+          R.filter(Boolean),
+          R.map((seq) => seq.getStore()),
+          R.unnest,
+          (sqlArray) => new Sql(sqlArray),
+        ),
+      ),
+    ),
+  )
+
+  const _getSqlSequenceChanges = () => _getSqlSequenceChangesFrom(_sequences)
+
   return Object.freeze({
     _getSqlCreateOrAlterTable,
     _getSqlConstraintChanges,
     _getSqlInsertSeeds,
+    _getSqlSequenceChanges,
     _getSchema,
     addSeeds,
   })