+++ src/sequence/parser.js
@@ -1,0 +1,19 @@
+/**
+ * Copyright (c) 2018-present Andrey Vereshchak
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+exports.dbSequence = (response) => {
+  if (response) {
+    const {
+      start_value: start,
+      minimum_value: min,
+      maximum_value: max,
+      cycle_option: cycle,
+      increment,
+    } = response
+    return { start, min, max, increment, cycle: cycle === 'YES' }
+  }
+}